Moving queue actions into its own controller
